sentiment_analysis/requirements.txt
azure-cognitiveservices-language-textanalytics==0.2.0
azure-common==1.1.23
certifi==2019.6.16
chardet==3.0.4
idna==2.8
isodate==0.6.0
jsonpickle==1.2
msrest==0.6.9
oauthlib==3.0.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.25.3
